+++ pkgs/test_core/lib/backend.dart
@@ -1,0 +1,8 @@
+// Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+export 'package:test_api/backend.dart'
+    show Metadata, PlatformSelector, Runtime, SuitePlatform;
+
+export 'src/runner/parse_metadata.dart' show parseMetadata;