+++ pkgs/test_core/lib/test_core.dart
@@ -10,9 +10,8 @@ export 'package:matcher/matcher.dart';
 import 'package:path/path.dart' as p;
 import 'package:pedantic/pedantic.dart';
 
+import 'package:test_api/backend.dart';
 import 'package:test_api/src/backend/invoker.dart'; // ignore: implementation_imports
-import 'package:test_api/src/backend/runtime.dart'; // ignore: implementation_imports
-import 'package:test_api/src/backend/suite_platform.dart'; // ignore: implementation_imports
 import 'package:test_api/src/frontend/timeout.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/declarer.dart'; // ignore: implementation_imports
 import 'package:test_api/src/utils.dart'; // ignore: implementation_imports