+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.4
+
+* Expose the `parseMetadata` function publicly through a new `backend.dart`
+  import, as well as re-exporting `package:test_api/backend.dart`.
+
 ## 0.2.3
 
 * Switch import for `IsolateChannel` for forwards compatibility with `2.0.0`.